requirements.txt
aiogram
yt-dlp
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
apscheduler
python-dotenv
aiohttp
aiofiles
